@@ -1,4 +1,5 @@
-import { makeStyles, useTheme } from "@material-ui/core";
+import { useTheme } from "@mui/material";
+
 import { SnackbarProvider } from "notistack";
 import {
   useContext,
@@ -34,12 +35,6 @@ export const sideBarNestedModes = {
   VIRTUAL_BACKGROUND: "VIRTUAL_BACKGROUND",
 };
 
-const useStyles = makeStyles((theme) => ({
-  container: {
-    backgroundColor: theme.palette.background.default,
-    color: theme.palette.text.primary,
-  },
-}));
 
 export const meetingLayouts = {
   SPOTLIGHT: "SPOTLIGHT",
@@ -166,7 +161,6 @@ export const MeetingAppProvider = ({
   const endCallContainerRef = useRef();
   const theme = useTheme();
 
-  const classes = useStyles();
   const [sideBarMode, setSideBarMode] = useState(null);
   const [sideBarNestedMode, setSideBarNestedMode] = useState(null);
   const [selectWebcamDeviceId, setSelectWebcamDeviceId] = useState(
@@ -415,7 +409,6 @@ export const MeetingAppProvider = ({
       }}
     >
       <SnackbarProvider
-        className={classes.container}
         autoHideDuration={5000}
         style={{
           backgroundColor:
@@ -423,7 +416,7 @@ export const MeetingAppProvider = ({
               ? theme.palette.darkTheme.seven
               : appTheme === appThemes.LIGHT
               ? theme.palette.lightTheme.main
-              : "",
+              : theme.palette.background.paper,
           color:
             appTheme === appThemes.LIGHT &&
             theme.palette.lightTheme.contrastText,
