@@ -1,12 +1,12 @@
 import {
   Box,
   ButtonBase,
-  makeStyles,
   Popover,
   Tooltip,
+  styled,
   useTheme,
-} from "@material-ui/core";
-import { Palette, FormatColorFill } from "@material-ui/icons";
+} from "@mui/material";
+import { Palette, FormatColorFill } from "@mui/icons-material";
 import React, { useEffect, useRef, useState } from "react";
 import {
   CircleFilledIcon,
@@ -27,39 +27,26 @@ import { appThemes, useMeetingAppContext } from "../../MeetingAppContextDef";
 import { SketchPicker } from "react-color";
 import UploadImageIcon from "../../icons/UploadImageIcon";
 
-const useStyles = makeStyles((theme) => ({
-  btnTool: {
-    padding: theme.spacing(1),
-    marginTop: theme.spacing(1 / 4),
-    marginBottom: theme.spacing(1 / 4),
-  },
-  iColorPicker: {
-    width: "24px",
-    height: "26px",
-    padding: "0px",
-    backgroundColor: "transparent",
-    border: "none",
-    cursor: "pointer",
-  },
-  popover: {
-    backgroundColor: "transparent",
-    boxShadow: "none",
-  },
-  paper: {
-    backgroundColor: "transparent",
-    boxShadow: "none",
-  },
-}));
+const CustomPopover = styled(Popover)`
+  & .MuiPaper-root {
+    background-color: transparent !important;
+    box-shadow: none !important;
+  }
+
+  & .MuiPopover-root {
+    background-color: transparent !important;
+    box-shadow: none !important;
+  }
+`;
 
 const ToolBarIcon = ({ Icon, onClick, title, isSelected }) => {
-  const classes = useStyles();
+  // const classes = useStyles();
   const theme = useTheme();
   const { appTheme } = useMeetingAppContext();
 
   return (
     <Tooltip title={title} arrow placement="right">
       <ButtonBase
-        className={classes.btnTool}
         color="inherit"
         style={{
           backgroundColor: isSelected
@@ -70,6 +57,9 @@ const ToolBarIcon = ({ Icon, onClick, title, isSelected }) => {
               : "#D5E8FF"
             : "",
           borderRadius: 6,
+          padding: theme.spacing(1),
+          marginTop: theme.spacing(1 / 4),
+          marginBottom: theme.spacing(1 / 4),
         }}
         onClick={onClick}
       >
@@ -95,18 +85,19 @@ const CustomColorPicker = ({
   setColor,
   setParentColor,
   changeCanvasBackgroundColor,
+  changeBrushColor,
   whiteboardToolbarWidth,
   whiteboardSpacing,
   Icon,
 }) => {
-  const classes = useStyles();
   const { appTheme } = useMeetingAppContext();
+  const theme = useTheme();
 
   return (
     <>
       <Tooltip title={title} arrow placement="right">
         <ButtonBase
-          className={classes.btnTool}
+          // className={classes.btnTool}
           color="inherit"
           component="span"
           onClick={(e) => {
@@ -116,6 +107,9 @@ const CustomColorPicker = ({
             backgroundColor:
               appTheme === appThemes.LIGHT ? "#596BFF33" : `#D5E8FF`,
             borderRadius: 6,
+            padding: theme.spacing(1),
+            marginTop: theme.spacing(1 / 4),
+            marginBottom: theme.spacing(1 / 4),
           }}
         >
           <Icon
@@ -127,11 +121,7 @@ const CustomColorPicker = ({
         </ButtonBase>
       </Tooltip>
 
-      <Popover
-        className={classes.popover}
-        classes={{
-          paper: classes.paper,
-        }}
+      <CustomPopover
         open={Boolean(colorPicker)}
         anchorEl={colorPicker}
         anchorOrigin={{
@@ -167,6 +157,9 @@ const CustomColorPicker = ({
               color={color}
               onChange={(ev) => {
                 setColor(ev.hex);
+                if (changeBrushColor) {
+                  changeBrushColor(ev.hex);
+                }
               }}
               onChangeComplete={(ev) => {
                 setParentColor(ev.hex);
@@ -177,15 +170,14 @@ const CustomColorPicker = ({
             />
           </Box>
         </Box>
-      </Popover>
+      </CustomPopover>
     </>
   );
 };
 
-const CustomImagePicker = ({ addImage }) => {
+const CustomImagePicker = ({ addImage, setTool }) => {
   const imageInputRef = useRef();
-
-  const classes = useStyles();
+  const theme = useTheme();
 
   return (
     <>
@@ -206,15 +198,18 @@ const CustomImagePicker = ({ addImage }) => {
             }}
           >
             <ButtonBase
-              className={classes.btnTool}
               color="inherit"
               style={{
                 borderRadius: 6,
                 position: "relative",
                 cursor: "pointer",
+                padding: theme.spacing(1),
+                marginTop: theme.spacing(1 / 4),
+                marginBottom: theme.spacing(1 / 4),
               }}
               onClick={() => {
                 imageInputRef.current.click();
+                setTool("select");
               }}
             >
               <input
@@ -244,6 +239,7 @@ const WBToolbar = ({
   downloadCanvas,
   clearCanvas,
   changeCanvasBackgroundColor,
+  changeBrushColor,
   undo,
   zoomOut,
   zoomIn,
@@ -256,7 +252,6 @@ const WBToolbar = ({
   whiteboardSpacing,
   addImage,
 }) => {
-  const classes = useStyles();
 
   const [color, setColor] = useState(parentColor);
   const [canvasBackgroundColor, setCanvasBackgroundColor] = useState(
@@ -332,7 +327,6 @@ const WBToolbar = ({
         />
         <>
           <ButtonBase
-            className={classes.btnTool}
             onMouseEnter={handlePopoverOpen}
             style={{
               backgroundColor:
@@ -345,6 +339,9 @@ const WBToolbar = ({
                     : "#D5E8FF"
                   : "",
               borderRadius: 6,
+              padding: theme.spacing(1),
+              marginTop: theme.spacing(1 / 4),
+              marginBottom: theme.spacing(1 / 4),
             }}
           >
             {tool === "square" ? (
@@ -384,11 +381,7 @@ const WBToolbar = ({
             )}
           </ButtonBase>
 
-          <Popover
-            className={classes.popover}
-            classes={{
-              paper: classes.paper,
-            }}
+          <CustomPopover
             open={open}
             anchorEl={anchorEl}
             anchorOrigin={{
@@ -461,7 +454,7 @@ const WBToolbar = ({
                 </Box>
               </Box>
             </Box>
-          </Popover>
+          </CustomPopover>
         </>
         <ToolBarIcon
           {...{
@@ -482,6 +475,7 @@ const WBToolbar = ({
             color: color,
             setParentColor: setParentColor,
             whiteboardToolbarWidth: whiteboardToolbarWidth,
+            changeBrushColor: changeBrushColor,
             whiteboardSpacing: whiteboardSpacing,
             Icon: FormatColorFill,
           }}
@@ -490,7 +484,10 @@ const WBToolbar = ({
         <ToolBarIcon
           {...{
             Icon: ZoomInIcon,
-            onClick: () => zoomIn(),
+            onClick: () => {
+              setTool("pan")
+              zoomIn();
+            },
             title: "Zoom In",
             isSelected: tool === "zoom",
             whiteboardToolbarWidth,
@@ -499,7 +496,10 @@ const WBToolbar = ({
         <ToolBarIcon
           {...{
             Icon: ZoomOutIcon,
-            onClick: () => zoomOut(),
+            onClick: () => {
+             setTool("pan")
+              zoomOut();
+            },
             title: "Zoom Out",
             isSelected: tool === "zoom",
             whiteboardToolbarWidth,
@@ -519,7 +519,7 @@ const WBToolbar = ({
             Icon: Palette,
           }}
         />
-        <CustomImagePicker addImage={addImage} />
+        <CustomImagePicker addImage={addImage} setTool={setTool} />
         <ToolBarIcon
           {...{
             Icon: UndoIcon,
